Fix socket interceptor and add unit test
The optimization of using a Stream instead of a Collection caused problems
with class not found and/or illegal access errors when using the lambda function
in the `Stream::forEach` call in the intercept method.
