+++ libs/agent-sm/agent/src/main/java/org/opensearch/javaagent/StackCallerProtectionDomainChainExtractor.java
@@ -10,13 +10,15 @@ package org.opensearch.javaagent;
 
 import java.lang.StackWalker.StackFrame;
 import java.security.ProtectionDomain;
+import java.util.Collection;
 import java.util.function.Function;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 /**
  * Stack Caller Chain Extractor
  */
-public final class StackCallerProtectionDomainChainExtractor implements Function<Stream<StackFrame>, Stream<ProtectionDomain>> {
+public final class StackCallerProtectionDomainChainExtractor implements Function<Stream<StackFrame>, Collection<ProtectionDomain>> {
     /**
      * Single instance of stateless class.
      */
@@ -32,10 +34,10 @@ public final class StackCallerProtectionDomainChainExtractor implements Function
      * @param frames stack frames
      */
     @Override
-    public Stream<ProtectionDomain> apply(Stream<StackFrame> frames) {
+    public Collection<ProtectionDomain> apply(Stream<StackFrame> frames) {
         return frames.map(StackFrame::getDeclaringClass)
             .map(Class::getProtectionDomain)
             .filter(pd -> pd.getCodeSource() != null) /* JDK */
-            .distinct();
+            .collect(Collectors.toSet());
     }
 }
