+++ libs/agent-sm/agent/src/test/java/org/opensearch/javaagent/AgentTests.java
@@ -8,21 +8,9 @@
 
 package org.opensearch.javaagent;
 
-import org.opensearch.javaagent.bootstrap.AgentPolicy;
-import org.junit.BeforeClass;
 import org.junit.Test;
 
-import java.security.Policy;
-import java.util.Set;
-
-public class AgentTests {
-    @SuppressWarnings("removal")
-    @BeforeClass
-    public static void setUp() {
-        AgentPolicy.setPolicy(new Policy() {
-        }, Set.of(), (caller, chain) -> caller.getName().equalsIgnoreCase("worker.org.gradle.process.internal.worker.GradleWorkerMain"));
-    }
-
+public class AgentTests extends AgentTestCase {
     @Test(expected = SecurityException.class)
     public void testSystemExitIsForbidden() {
         System.exit(0);
